fix bugs

+++ src/motor/Motor.h
@@ -2,18 +2,19 @@
 #define Motor_h
 
 #include <Arduino.h>
+#include <stdint.h>
 #include "AbstractMotor.h"
 #include "../vendor/AVR_PWM.h"
 
 class Motor : public AbstractMotor {
     private:
-        int pwmPin;
-        int directionPin;
+        uint8_t pwmPin;
+        uint8_t directionPin;
         int pwmFrequency;
         AVR_PWM* pwmControl;
 
     public:
-        Motor(int pwmPin, int directionPin, int pwmFrequency);
+        Motor(uint8_t pwmPin, uint8_t directionPin, int pwmFrequency);
         ~Motor();
         void setSpeed(int speed);
 };
