+++ src/motor/SteeringMotor.h
@@ -2,16 +2,17 @@
 #define SteeringMotor_h
 
 #include <Arduino.h>
+#include <stdint.h>
 #include "AbstractMotor.h"
 
 class SteeringMotor : public AbstractMotor {
     private:
-        int pwmPin;
-        int directionPin;
-        int relayPin;
+        uint8_t pwmPin;
+        uint8_t directionPin;
+        uint8_t relayPin;
 
     public:
-        SteeringMotor(int pwmPin, int directionPin, int relayPin);
+        SteeringMotor(uint8_t pwmPin, uint8_t directionPin, uint8_t relayPin);
         void setSpeed(int speed);
 };
